Manually commit to the cache db during the eval
This is a dirty hack to make caching temporarily work.

The issue seems to be that the cache object is never deleted, which in
turn means that we never commit to the db (as it happens in the object's
destructor), so the cache doesn't persist between two Nix calls.
So we manually commit to the db each time. This is both slower and ugly,
but at least it works for the moment